Improve outline hierarchy
By providing the container name and changing the symbols ranges to
contain the entire nodes ranges not just the name of the symbol.

The last part has the unfortunate consequence that when you go to a
definition it will no longer go to the exact word but instead the line
where the symbol is defined.