Replace complex tree sitter helper
The helper was originally introduced to make the detection of a word at a given point
more accurate (this is used for completion, hovering, etc) and fixes some issues with the tree sitter grammar.

It turns out this is really not needed and introduces a bunch of issues.

Related to https://github.com/bash-lsp/bash-language-server/pull/192